scanner/requirements.txt
requests
psycopg2-binary
python-Wappalyzer
python-whois
beautifulsoup4
tldextract
python-dotenv

